+++ packages/nimble-components/src/patterns/severity/styles.ts
@@ -1,10 +1,9 @@
 import { cssPartial } from '@ni/fast-element';
 import {
     failColor,
-    errorTextFontLineHeight,
     errorTextFont,
     buttonLabelFontColor,
-    warningFontColor
+    warningFontColor,
 } from '../../theme-provider/design-tokens';
 
 // These styles end up inside a @layer block so must use the
@@ -14,16 +13,14 @@ export const styles = cssPartial`
         display: none;
         font: ${errorTextFont};
         color: ${buttonLabelFontColor};
-        width: 100%;
         position: absolute;
-        ${'' /* The -2px modifier of the bottom position is to intentionally have the severity text slightly overlap the control by 2px */}
-        bottom: calc(-1 * (${errorTextFontLineHeight} - 2px));
-        left: 0px;
+        padding: 0;
         overflow: hidden;
         text-overflow: ellipsis;
         white-space: nowrap;
+        width: 100%;
     }
-    
+
     :host([severity="error"]) .severity-text {
         display: block;
         color: ${failColor};