+++ packages/react-workspace/ok-react/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/ok-react
 
-<!-- This log was last generated on Thu, 12 Mar 2026 16:41:20 GMT and should not be manually modified. -->
+<!-- This log was last generated on Wed, 18 Mar 2026 21:46:50 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.6.2
+
+Wed, 18 Mar 2026 21:46:50 GMT
+
+### Patches
+
+- Bump @ni/ok-components to v0.3.2
+
+## 0.6.1
+
+Mon, 16 Mar 2026 19:23:09 GMT
+
+### Patches
+
+- Bump @ni/ok-components to v0.3.1
 
 ## 0.6.0